+++ src/cengine/chunk.h
@@ -38,7 +38,7 @@ typedef struct chunk{
   struct chunk *nz;
 } chunk_t;
 
-unsigned int block_index(uint8_t x, uint8_t y, uint8_t z);
+uint16_t block_index(uint8_t x, uint8_t y, uint8_t z);
 chunk_t chunk_init(int x, int y, int z);
 void chunk_free(chunk_t *chunk);
 unsigned char chunk_update(chunk_t *chunk);
